pyproject.toml
[tool.poetry]
name = "telegram-battery-bot"
version = "0.1.0"
description = "Telegram bot that notifies when battery is low"
authors = ["Tomasz Walotek <tomasz.walotek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^12.2"
psutil = "^5.6"
docopt = "^0.6.2"

[tool.poetry.dev-dependencies]
mypy = "^0.750.0"
black = {version = "^18.3-alpha.0", allows-prereleases = true}

[tool.poetry.scripts]
batbottg = 'telegram_battery_bot:run_bot'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
